dlt/requirements.txt
dlt
google-cloud-storage
pyyaml
requests
pandas
python-dotenv
psycopg2-binary